+++ inst/examples/ex-scale_linetype_stata.R
@@ -1,6 +1,8 @@
 
 
-library("tidyverse")
+require("dplyr")
+require("tidyr")
+require("ggplot2")
 rescale01 <- function(x) {(x - min(x)) / diff(range(x))}
 gather(economics, variable, value, -date) %>%
   group_by(variable) %>%